@@ -19,16 +19,15 @@ classifiers =
     Programming Language :: Python
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
 
 [options]
 packages = find:
 install_requires =
     requests-oauthlib
-python_requires = >=3.6
+python_requires = >=3.8
 include_package_data = True
 zip_safe = False
 
@@ -41,16 +40,6 @@ tests =
 exclude =
     tests
     testapp
-
-[flake8]
-exclude = venv,build,docs,.tox,migrations
-max-complexity = 10
-max-line-length = 88
-
-[isort]
-profile = black
-combine_as_imports = True
-lines_after_imports = 2
 
 [coverage:run]
 branch = True
