@@ -3,7 +3,7 @@ from django.core.cache import cache
 from requests_oauthlib import OAuth1Session
 
 
-class TwitterOAuthClient(object):
+class TwitterOAuthClient:
     authorization_base_url = "https://api.twitter.com/oauth/authenticate"
     client_id = settings.TWITTER_CLIENT_ID
     client_secret = settings.TWITTER_CLIENT_SECRET
